% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.spearman.R
\name{cor.spearman}
\alias{cor.spearman}
\title{Spearman rank correlation}
\usage{
cor.spearman(x,y)
}
\arguments{
\item{x}{:a numeric variable}

\item{y}{:a numeric variable}
}
\description{
Calculate Spearman Rank Correlation, which is the nonparametric version of the Pearson product-moment correlation.
}
\examples{
 cor.spearman(mtcars[,1], mtcars[,3])
}
