% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.annu.fv.R
\name{cv.annu.fv}
\alias{cv.annu.fv}
\title{Calculate future value of annuity}
\usage{
cv.annu.fv(pmt,i,n,type = 0)
}
\arguments{
\item{pmt}{:the equal amount of payment of each period}

\item{i}{:interest rate according to the period}

\item{n}{:number of periods}

\item{type}{:type = 0 for ordinary annuity, type = 1 for annuity due}
}
\description{
Calculate future value of an ordinary annuity or an annuity due.
}
\examples{
cv.annu.fv(100,0.0248,10,0)
}
