% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.annu.pv.R
\name{cv.annu.pv}
\alias{cv.annu.pv}
\title{Calculate present value of annuity}
\usage{
cv.annu.pv(pmt,i,n,k)
}
\arguments{
\item{pmt}{:the equal amount of payment of each period}

\item{i}{:interest rate according to the period}

\item{n}{:number of periods}

\item{k}{:number of periods deferred until first payment}
}
\description{
Calculate present value of an ordinary annuity or an annuity due.
}
\examples{
cv.annu.pv(100,0.0248,10,4)
}
