% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.axp.R
\name{cv.axp}
\alias{cv.axp}
\title{Create logarithm with a random base}
\usage{
cv.axp(dataframe, var, n, range)
}
\arguments{
\item{dataframe}{:a data frame}

\item{var}{:the variable selected}

\item{n}{:number of new variables created}

\item{range}{:the range of base}
}
\description{
Create a new variable with the base of a random number and power of the selected variable
}
\examples{
cv.axp(mtcars,"wt",5,c(1, 2))
}
