% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.bondprice.R
\name{cv.bondprice}
\alias{cv.bondprice}
\title{Calculate the plain vanilla bond price}
\usage{
cv.bondprice(par,c,yield,n,m)
}
\arguments{
\item{par}{:the face value of the bond}

\item{c}{:the annual coupon rate of the bond}

\item{yield}{:the annual yield to maturity of a bond}

\item{n}{:number of years}

\item{m}{:compounding period in a year}
}
\description{
Calculate the plain vanilla bond price
}
\examples{
cv.bondprice(1000,0.0248,0.0248,10,2)
}
