% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.lag.R
\name{cv.lag}
\alias{cv.lag}
\title{Create a lag variable}
\usage{
cv.lag(x,n)
}
\arguments{
\item{x}{:a vector}

\item{n}{:number of lag periods}
}
\description{
Create a lag variable, with a choice of lag periods. The lag variable can be used
to test lag effects between variables.
}
\examples{
cv.lag(mtcars[,2],3)
data.frame(mtcars,cv.lag(mtcars[,3], 1))
}
