% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.lead.R
\name{cv.lead}
\alias{cv.lead}
\title{Create a lead variable}
\usage{
cv.lead(x,n)
}
\arguments{
\item{x}{:a vector}

\item{n}{:number of lead periods}
}
\description{
Create a lead variable, with a choice of lead periods. The lead variable can be used
to test lead effects between variables.
}
\examples{
cv.lead(mtcars[,2],3)
data.frame(mtcars,cv.lead(mtcars[,3], 3))
}
