% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.logs.R
\name{cv.logs}
\alias{cv.logs}
\title{Create logarithm with a random base}
\usage{
cv.logs(dataframe, var, n, range)
}
\arguments{
\item{dataframe}{:a data frame}

\item{var}{:the variable selected}

\item{n}{:number of new variables created}

\item{range}{:the range of base}
}
\description{
Create a new variable that is the logarithm of the selected variable with the base of a random number
}
\examples{
cv.logs(mtcars,"wt",5,c(1, 2))
}
