% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.pctcng.R
\name{cv.pctcng}
\alias{cv.pctcng}
\title{Calculating rate of return of a vector}
\usage{
cv.pctcng(x,n)
}
\arguments{
\item{x}{:a numeric vector}

\item{n}{: number of lag periods}
}
\description{
Calculating the percentage change of a time series vector for further analysis, including calculating beta of companies, plotting to see the trend of the stock for technical analysis.
}
\examples{
cv.pctcng(mtcars[,1],1)
}
