% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.powers.R
\name{cv.powers}
\alias{cv.powers}
\title{Create nth power variable}
\usage{
cv.powers(dataframe, var, n, range)
}
\arguments{
\item{dataframe}{:a data frame}

\item{var}{:the variable selected}

\item{n}{:number of new variables created}

\item{range}{:the range of power}
}
\description{
Create a new variable that is the nth power of the selected variable
}
\examples{
cv.powers(mtcars,"wt",5,c(1, 2))
}
