% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.corm.R
\name{ds.corm}
\alias{ds.corm}
\title{Correlation matrix}
\usage{
ds.corm(x,n)
}
\arguments{
\item{x}{:a data frame}

\item{n}{:number of decimal points}
}
\description{
Calculating the correlation matrix of a data frame and return in a data frame object
}
\examples{
ds.corm(mtcars,3)
}
