% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.kurt.R
\name{ds.kurt}
\alias{ds.kurt}
\title{Calculating kurtosis for numeric data.}
\usage{
ds.kurt(x)
}
\arguments{
\item{x}{:a numeric variable}
}
\description{
Kurtosis
}
\examples{
ds.kurt(mtcars[,2])
}
