% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.skew.R
\name{ds.skew}
\alias{ds.skew}
\title{Calculating skewness for numeric data}
\usage{
ds.skew(x, type = 3)
}
\arguments{
\item{x}{:a numeric variable}

\item{type}{:type = 1 for mode skewness; type = 2 for median skewness; type = 3 for mean skewness}
}
\description{
Calculating Pearson's skewness in three types: mode, median, and mean.
}
\examples{
ds.skew(mtcars[,1])
}
