% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.summ.R
\name{ds.summ}
\alias{ds.summ}
\title{Descriptive statistics of a data frame}
\usage{
ds.summ(x,n)
}
\arguments{
\item{x}{:a data frame}

\item{n}{:number of decimal points rounded}
}
\description{
Calculating the descriptive statistics of a data frame and exporting in a data frame.
The report data frame contains: number of observations, maximum value, minimum value, mean, median,
mode, variance, standard deviation, skewness and kurtosis.
}
\examples{
ds.summ(mtcars,3)
}
