% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.reinstall.R
\name{pkg.reinstall}
\alias{pkg.reinstall}
\title{Reinstalled packages after updating R}
\usage{
pkg.reinstall(path)
}
\arguments{
\item{path}{:a path that contains saved RDA file}
}
\description{
Reinstall packages from pre-saved RDA file, using pkg.save(). If no path is given, the function will try to look in the current working directory.
}
\examples{
path <- getwd()
pkg.save(path)
pkg.reinstall(path)
}
