% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.save.R
\name{pkg.save}
\alias{pkg.save}
\title{Save installed packages}
\usage{
pkg.save(path)
}
\arguments{
\item{path}{:a path to save installed package file}
}
\description{
Save installed packages to an RDA file in a given directory. If no path is given, the file will be saved in the current working directory.
}
\examples{
path <- getwd()
pkg.save(path)
}
