% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.2ts.R
\name{pl.2ts}
\alias{pl.2ts}
\title{Time series plot for two variables}
\usage{
pl.2ts(ts1,ts2,title)
}
\arguments{
\item{ts1}{:time series variable one}

\item{ts2}{:time series variable two}

\item{title}{:title for the plot}
}
\description{
Plotting two time series in one plot, with title.
}
\examples{
DAX <- EuStockMarkets[,1]
FTSE <- EuStockMarkets[,4]
pl.2ts(DAX,FTSE, "Times Series Plot of DAX and FTSE")
}
