% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.3smoothtxtgg.R
\name{pl.3smoothtxtgg}
\alias{pl.3smoothtxtgg}
\title{Scatter smooth plot with text overlay using ggplot2}
\usage{
pl.3smoothtxtgg(x,y,txt,size,title,xlab,ylab)
}
\arguments{
\item{x}{:a numeric vector}

\item{y}{:a numeric vector}

\item{txt}{:a vector used as labels}

\item{size}{:text size, which default is set as 3}

\item{title}{:graph title}

\item{xlab}{:x-axis label}

\item{ylab}{:y-axis label}
}
\description{
Generate a scatter plot with text overlay, with a smooth curve fitted by loess.
}
\examples{
pl.3smoothtxtgg(mtcars[,1], mtcars[,3], row.names(mtcars), 3, "MPG v. DISP","mpg","disp")
}
