% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.3txtgg.R
\name{pl.3txtgg}
\alias{pl.3txtgg}
\title{Scatter plot with text overlay with ggplot2}
\usage{
pl.3txtgg(x,y,txt,size,title,xlab,ylab)
}
\arguments{
\item{x}{:a numeric vector}

\item{y}{:a numeric vector}

\item{txt}{:a vector used as labels}

\item{size}{:text size, which default is set as 3}

\item{title}{:title of the graph}

\item{xlab}{:x-axis label}

\item{ylab}{:y-axis label}
}
\description{
Generate a scatter plot with text overlay with ggplot2. This plot is to better show the effect of
the text variable in the domain of x and y variable.
}
\examples{
pl.3txtgg(mtcars[,1], mtcars[,3], row.names(mtcars), 3,"mpg v. cyl", "mpg", "cyl")
}
