% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.coplot.R
\name{pl.coplot}
\alias{pl.coplot}
\title{Scatter plot of x and y divided by z}
\usage{
pl.coplot(x,y,z,varN)
}
\arguments{
\item{x}{:x-axis value}

\item{y}{:y-axis value}

\item{z}{:classification variable used to condition plots based on ascending values of z}

\item{varN}{:variable name of z}
}
\description{
Generate 4 scatter plots of x and y divided by variable z, with a fitted line using a robust linear regression method.
}
\examples{
pl.coplot(mtcars[,1], mtcars[,3], mtcars[,4], "hp")
}
