% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.hist.R
\name{pl.hist}
\alias{pl.hist}
\title{Plot histograms  for a data frame}
\usage{
pl.hist(x, l = 1)
}
\arguments{
\item{x}{:a data frame}

\item{l}{: the beginning label number in the title (default set to 1)}
}
\description{
Plotting histograms for a data frame, with titles and label numbers.
}
\examples{
pl.hist(mtcars,1)
}
