% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.histgg.R
\name{pl.histgg}
\alias{pl.histgg}
\title{Plot histograms for a data frame with ggplot2}
\usage{
pl.histgg(x,l,bin)
}
\arguments{
\item{x}{:a data frame}

\item{l}{:the beginning label number in the title (default set to 1)}

\item{bin}{:bin width of histogram (default set to 30)}
}
\description{
Plotting histograms for a data frame with 4 per page, with titles and label numbers automatically generated.
}
\examples{
pl.histgg(as.data.frame(EuStockMarkets),1)
}
