% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.hsd.R
\name{pl.hsd}
\alias{pl.hsd}
\title{Plot histogram with density line for a data frame}
\usage{
pl.hsd(dataframe,l)
}
\arguments{
\item{dataframe}{:a data frame}

\item{l}{: the beginning label number in the title (default set to 1)}
}
\description{
Plotting histogram with density for a data frame, with titles and label numbers.
}
\examples{
pl.hsd(mtcars,1)
}
