% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.hsdgg.R
\name{pl.hsdgg}
\alias{pl.hsdgg}
\title{Plot histograms for a data frame with ggplot2}
\usage{
pl.hsdgg(x,l,bin)
}
\arguments{
\item{x}{:a data frame}

\item{l}{:the beginning label number in the title (default set to 1)}

\item{bin}{:bin width of the graph}
}
\description{
Plotting histograms for a data frame with 4 per page, with titles and label numbers automatically generated.
}
\examples{
pl.hsdgg(as.data.frame(EuStockMarkets),1,100)
}
