% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.mv.R
\name{pl.mv}
\alias{pl.mv}
\title{Plot mean-variance simulation result}
\usage{
pl.mv(port)
}
\arguments{
\item{port}{:portfolio simulation result from pt.mv()}
}
\description{
This function is used to plot the result of portfolio simulation by pt.mv().
}
\examples{
set.seed(1)
rtn <- data.frame(runif(120,-1,1),runif(120,-1,1),runif(120,-1,1),runif(120,-1,1))
names(rtn) <- c("asset1","asset2","asset3","asset4")
portfolio <- pt.hismv(rtn,1000,0)
pl.mv(portfolio)
}
