% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.sgg.R
\name{pl.sgg}
\alias{pl.sgg}
\title{Plot scatter plots for a data frame using ggplot2}
\usage{
pl.sgg(x,dependent,l)
}
\arguments{
\item{x}{:a data frame, which includes the dependent variable}

\item{dependent}{:the dependent variable for scatter plot}

\item{l}{: the beginning label number in the title (default set to 1)}
}
\description{
Plotting scatter plots for a data frame using ggplot2, with titles and label numbers. The output will be 4 graphs per page.
}
\examples{
pl.sgg(mtcars,"mpg",1)
}
