% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.sm.R
\name{pl.sm}
\alias{pl.sm}
\title{Plot scatter smooth plots for a data frame}
\usage{
pl.sm(x,dependent,l)
}
\arguments{
\item{x}{:a data frame, which includes the dependent variable}

\item{dependent}{:the dependent variable for scatter smooth plots}

\item{l}{: the beginning label number in the title (default set to 1)}
}
\description{
Plotting scatter smooth plots for a data frame, with titles and label numbers.
}
\examples{
pl.sm(mtcars,"mpg",1)
}
