% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.smgg.R
\name{pl.smgg}
\alias{pl.smgg}
\title{Plot scatter plots with smooth line for a data frame using ggplot2}
\usage{
pl.smgg(x,dependent,l,mtd)
}
\arguments{
\item{x}{:a data frame, which includes the dependent variable}

\item{dependent}{:the dependent variable for scatter plot}

\item{l}{:the beginning label number in the title (default set to 1)}

\item{mtd}{:sommthing method to use, accepts either a character vector or a function, e.g. MASS::rlm, base::lm, base::loess, mgcv::gam}
}
\description{
Plotting scatter plots for a data frame using ggplot2, with titles and label numbers. A smooth line will be added using a chosen method. The output will be 4 graphs per page.
}
\examples{
pl.smgg(mtcars,"mpg",1,lm)
pl.smgg(mtcars,"mpg",1,loess)
}
