% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.ts.R
\name{pl.ts}
\alias{pl.ts}
\title{Plot time series plots  for a data frame}
\usage{
pl.ts(x, l = 1)
}
\arguments{
\item{x}{:a data frame}

\item{l}{: the beginning label number in the title (default set to 1)}
}
\description{
Plotting time series plots for a data frame, with titles and label numbers.
}
\examples{
pl.ts(mtcars,1)
}
