% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.tss.R
\name{pl.tss}
\alias{pl.tss}
\title{Time series plot with multiple variables}
\usage{
pl.tss(dataframe,ylb,title)
}
\arguments{
\item{dataframe}{:a data frame}

\item{ylb}{:y-axis label}

\item{title}{:plot title}
}
\description{
This function will return a time series plot with up to 6 variables, each with different line type.
}
\examples{
pl.tss(EuStockMarkets,"Price","Daily Closing Prices of Major European Stock Indices")
}
