% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.alpha.R
\name{pt.alpha}
\alias{pt.alpha}
\title{Stock return alpha}
\usage{
pt.alpha(ar,br)
}
\arguments{
\item{ar}{:a vector of a risk asset return}

\item{br}{:a vector of benchmark return}
}
\description{
Alpha is the intercept of a fitted line when dependent variable is the benchmark return
and independent variable is a asset return of the same period. It is a measure of the active return
on an investment. Alpha, along with beta, is one of the two key coefficients in the CAPM used modern portfolio theory.
}
\examples{
brtn <- runif(100, -1, 1)
artn <- runif(100, -1, 1)
pt.alpha(artn,brtn)
}
