% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.annexrtn.R
\name{pt.annexrtn}
\alias{pt.annexrtn}
\title{Annualized excess return}
\usage{
pt.annexrtn(ar,br)
}
\arguments{
\item{ar}{:a vector of a risk asset return}

\item{br}{:a vector of benchmark return}
}
\description{
Annualized excess return is the difference between the annualized and cumulative
return of the two series. Usually, one series are portfolio returns and the other is a benchmark
returns.
}
\examples{
artn <- runif(100, -1, 1)
brtn <- runif(100, -1, 1)
pt.annexrtn(artn, brtn)
}
