% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.annrtn.R
\name{pt.annrtn}
\alias{pt.annrtn}
\title{Annualized return}
\usage{
pt.annrtn(r,n)
}
\arguments{
\item{r}{:annual returns}

\item{n}{:number of years}
}
\description{
This function takes a series of annual returns and calculate the annualized return.
}
\examples{
r <- runif(100,-1,1) # generate random number to simulate returns
annualizedreturn <- pt.annrtn(r,100)
}
