% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.annsd.R
\name{pt.annsd}
\alias{pt.annsd}
\title{Annualized standard deviation}
\usage{
pt.annsd(r,n)
}
\arguments{
\item{r}{:a vector of a risk asset return}

\item{n}{:number of periods in a year}
}
\description{
The annualized standard deviation is the standard deviation multiplied by the
square root of the number of periods in one year.
}
\examples{
rtn <- runif(30, -1, 1)
n <- 30
pt.annsd(rtn,n)
}
