% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.beta.R
\name{pt.beta}
\alias{pt.beta}
\title{Stock return beta}
\usage{
pt.beta(ar,br)
}
\arguments{
\item{ar}{:a vector of a risk asset return}

\item{br}{:a vector of benchmark return}
}
\description{
Beta is the slope of a fitted line when dependent variable is the benchmark return
and independent variable is an asset return of the same period. It is a measure the risk arising
from exposure to general market movements.
}
\examples{
brtn <- runif(100, -1, 1)
artn <- runif(100, -1, 1)
pt.beta(artn, brtn)
}
