% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.bias.R
\name{pt.bias}
\alias{pt.bias}
\title{Bias ratio}
\usage{
pt.bias(r)
}
\arguments{
\item{r}{:a vector of a risk asset return}
}
\description{
The bias ratio is an indicator used in finance analyze the returns of a portfolio, and
in performing due diligence.
}
\examples{
r <- runif(100,0,1) # generate random number to simulate returns
pt.bias(r)
}
