% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.cmrtn.R
\name{pt.cmrtn}
\alias{pt.cmrtn}
\title{Cumulative return}
\usage{
pt.cmrtn(r)
}
\arguments{
\item{r}{:a vector of periodic returns}
}
\description{
Cumulative return is the compounded return in a given period.
}
\examples{
rt <- runif(12,-1,1) # generate random number to simulate returns
pt.cmrtn(rt)
}
