% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.dalpha.R
\name{pt.dalpha}
\alias{pt.dalpha}
\title{Dual-alpha}
\usage{
pt.dalpha(ar,mr,rf)
}
\arguments{
\item{ar}{:a vector of a risk asset return}

\item{mr}{:a vector of market return}

\item{rf}{:risk free rate}
}
\description{
Dual-alpha method is to divide market alpha into downside beta and upside alpha. The principle
behind is that upside and downside alphas are not the same.
}
\examples{
artn <- runif(24,0,1) # generate random number to simulate returns
mrtn <- runif(24,-1,1)
pt.dalpha(artn,mrtn,0.024)
}
