% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.info.R
\name{pt.info}
\alias{pt.info}
\title{Information ratio}
\usage{
pt.info(ar,br,n)
}
\arguments{
\item{ar}{:a vector of a risk asset return}

\item{br}{:a vector of benchmark return}

\item{n}{:number of years}
}
\description{
The information ratio of asset's returns versus benchmark returns, is the
quotient of the annualized excess return and the annualized standard deviation of the excess
return.
}
\examples{
brtn <- runif(100, -1, 1)
artn <- runif(100, 0, 1)
pt.info(artn,brtn,100)
}
