% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.jalpha.R
\name{pt.jalpha}
\alias{pt.jalpha}
\title{Jensen's alpha}
\usage{
pt.jalpha(pr,mr,rf,beta)
}
\arguments{
\item{pr}{:portfolio return}

\item{mr}{:market return}

\item{rf}{:risk free rate}

\item{beta}{:portfolio beta}
}
\description{
Jensen's alpha is a financial statistic used to quantify the abnormal return of
a security or portfolio over the theoretical expected return. Unlike, standard alpha, it uses
theoretical performance return instead of a market return.
}
\examples{
prtn <- runif(24, -1, 1)
mrtn <- runif(24, -1, 1)
rf <- 0.024
pt.jalpha(mean(prtn), mean(mrtn), rf, pt.beta(prtn,mrtn))

}
