% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.m2.R
\name{pt.m2}
\alias{pt.m2}
\title{Modigliani risk-adjusted performance}
\usage{
pt.m2(pr,br,rf)
}
\arguments{
\item{pr}{:portfolio return}

\item{br}{:benchmark return}

\item{rf}{:risk free rate}
}
\description{
Modigliani risk-adjusted performance is a financial measure of risk-adjusted returns
of a portfolio. It measures the returns of the portfolio after adjusting it relative to some benchmark.
}
\examples{
prtn <- runif(12,-1,1)
brtn <- runif(12,-1,1)
rf <- 0.024
pt.m2(prtn,brtn,rf)
}
