% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.probloss.R
\name{pt.probloss}
\alias{pt.probloss}
\title{Probability of loss}
\usage{
pt.probloss(r,p)
}
\arguments{
\item{r}{:a vector of periodic returns}

\item{p}{:target return}
}
\description{
This function give the probability of loss of given asset returns.
}
\examples{
rt <- runif(12,-1,1) # generate random number to simulate returns
pt.probloss(rt,0)
pt.probloss(rt,0.05)
}
