% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.roy.R
\name{pt.roy}
\alias{pt.roy}
\title{Roy's safety-first criterion}
\usage{
pt.roy(r,mar)
}
\arguments{
\item{r}{:a vector of a risk asset return}

\item{mar}{:minimum acceptable return}
}
\description{
Roy's safety-first criterion is a risk management technique that allows to choose
a portfolio based on the criterion that the probability of the portfolio's return falling below
a minimum desired threshold is minimized.
}
\examples{
r <- runif(100,0,1) # generate random number to simulate returns
pt.roy(r,0.024)
}
