% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.sdexrtn.R
\name{pt.sdexrtn}
\alias{pt.sdexrtn}
\title{Standard deviation of excess return}
\usage{
pt.sdexrtn(ar,br)
}
\arguments{
\item{ar}{:a vector of a risk asset return}

\item{br}{:a vector of benchmark return}
}
\description{
The standard deviation of excess return is simply the standard deviation of the asset
return over the benchmark return.
}
\examples{
artn <- runif(12, -1, 1)
brtn <- runif(12,-1,1)
pt.sdexrtn(artn,brtn)
}
