% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.semivar.R
\name{pt.semivar}
\alias{pt.semivar}
\title{Semivariance of loss}
\usage{
pt.semivar(r,p)
}
\arguments{
\item{r}{:a vector of periodic returns}

\item{p}{:target return}
}
\description{
This function give the semivariance of a losing scenario.
}
\examples{
rt <- runif(12,-1,1) # generate random number to simulate returns
pt.semivar(rt,0)
pt.semivar(rt,0.03)
}
