% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.sortino.R
\name{pt.sortino}
\alias{pt.sortino}
\title{Sortino ratio}
\usage{
pt.sortino(r,p,n,rf)
}
\arguments{
\item{r}{:a vector of a risk asset return}

\item{p}{:target return, aka minimum acceptable return(MAR)}

\item{n}{:number of years of asset return, used to calculate annualized return}

\item{rf}{:risk free rate}
}
\description{
The Sortino ratio is an analog to the sharp ratio, with standard deviation replaced
by the downside deviation.
}
\examples{
rtn <- runif(12, -1, 1)
pt.sortino(rtn,0.3,1,0.024)
}
