% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.te.R
\name{pt.te}
\alias{pt.te}
\title{Tracking error}
\usage{
pt.te(pr,br)
}
\arguments{
\item{pr}{:portfolio return}

\item{br}{:benchmark return}
}
\description{
Tracking error, in finance, is a measure of risk in a portfolio that is due to active management
decisions made by the manager. It indicates how closely the portfolio follows the benchmark of choosing.
}
\examples{
prtn <- runif(12,-1,1)
brtn <- runif(12,-1,1)
pt.te(prtn,brtn)
}
