% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.treynor.R
\name{pt.treynor}
\alias{pt.treynor}
\title{Treynor ratio}
\usage{
pt.treynor(ar,br,n,rf)
}
\arguments{
\item{ar}{:a vector of a risk asset return}

\item{br}{:a vector of benchmark return}

\item{n}{:number of years of asset return, used to calculate annualized return}

\item{rf}{:risk free rate}
}
\description{
The Treynor ratio is an analog to the sharp ratio, with standard deviation replaced
by the asset beta to benchmark.
}
\examples{
rtn <- runif(24, -1, 1)
brtn <- runif(24,-1,1)
pt.treynor(rtn,brtn,2,0.024)
}
