% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.udrtn.R
\name{pt.udrtn}
\alias{pt.udrtn}
\title{Average up and down returns}
\usage{
pt.udrtn(r)
}
\arguments{
\item{r}{:a vector of periodic returns}
}
\description{
This function calculates the average up and down returns from a series of returns.
}
\examples{
r <- runif(100,-1,1) # generate random number to simulate returns
pt.udrtn(r)
}
