% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.updwcap.R
\name{pt.updwcap}
\alias{pt.updwcap}
\title{Up and down capture}
\usage{
pt.updwcap(ar,br,n)
}
\arguments{
\item{ar}{:a vector of a risk asset return}

\item{br}{:a vector of benchmark return}

\item{n}{:number of years of asset return, used to calculate annualized return}
}
\description{
The up and down capture is a measure of how an asset was able to improve on benchmark returns
or how it underperforms over the benchmark.
}
\examples{
artn <- runif(12, -1, 1)
brtn <- runif(12,-1,1)
pt.updwcap(artn,brtn,1)
}
