% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.aic.R
\name{reg.aic}
\alias{reg.aic}
\title{AIC for lm.fit}
\usage{
reg.aic(fit,w)
}
\arguments{
\item{fit}{:the outcome of lm.fit}

\item{w}{:wright}
}
\description{
Calculate AIC for the outcome of AIC. This function is built for reg.linreg for higher efficiency only.
It can't be used for calculating AIC in general operation.
}
\examples{
X <- as.matrix(cbind(1,EuStockMarkets[,1:2])) # create the design matrix
Y <- as.data.frame(EuStockMarkets)$FTSE
fit <- lm.fit(x = X, y = Y)
w <- rep(1,length(Y))
reg.aic(fit,w)
}
