% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.dof.R
\name{reg.dof}
\alias{reg.dof}
\title{Degree of freedom for lim.fit}
\usage{
reg.dof(fit)
}
\arguments{
\item{fit}{:outcome of lm.f}
}
\description{
Calculate degree of freedom for the outcome of lm.fit(). This function is built for reg.linreg for higher efficiency only.
It can't be used for calculating degree of freedom in general operation.
}
\examples{
X <- as.matrix(cbind(1,EuStockMarkets[,1:2])) # create the design matrix
Y <- as.data.frame(EuStockMarkets)$FTSE
fit <- lm.fit(x = X, y = Y)
reg.dof(fit)
}
