% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.dw.R
\name{reg.dw}
\alias{reg.dw}
\title{Durbin-Watson Test}
\usage{
reg.dw(fit)
}
\arguments{
\item{fit}{:a lm object}
}
\description{
Performs the Durbin-Watson Test for a regression model
}
\examples{
fit <- lm(mpg~wt, mtcars, na.action = na.omit)
reg.dw(fit)
}
