% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.model.R
\name{reg.model}
\alias{reg.model}
\title{Linear model generator}
\usage{
reg.model(dataframe,dependent)
}
\arguments{
\item{dataframe}{:a data frame}

\item{dependent}{: dependent variable}
}
\description{
This function will take a data frame and generate all the combinations of linear model
}
\examples{
reg.model(mtcars,"mpg")
}
