% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.logtb.R
\name{tr.logtb}
\alias{tr.logtb}
\title{Logistic function}
\usage{
tr.logtb(x,top,bot,a,b)
}
\arguments{
\item{x}{:a vector}

\item{top}{:higher level y asymptote}

\item{bot}{:lower level y asymptote}

\item{a}{:a number to control top acceleration of the curve}

\item{b}{:a number to control bottom acceleration of the curve}
}
\description{
Generate logistic series, with set top and bottom value and acceleration.
}
\examples{
tr.logtb(seq(-3, 3, 0.1), 1, 0.4, -3, 3)
}
