% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.nd.R
\name{tr.nd}
\alias{tr.nd}
\title{Normal density function}
\usage{
tr.nd(x,mu,sig)
}
\arguments{
\item{x}{:x value}

\item{mu}{:mean value}

\item{sig}{:standard deviation}
}
\description{
Calculate normal density function value at x with a mean of mu and standard deviation of sig.
}
\examples{
 tr.nd(seq(-3, 3, 0.1), 0, 1)
}
