% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.unli.R
\name{tr.unli}
\alias{tr.unli}
\title{Unit normal loss integral}
\usage{
tr.unli(x,disc,disp)
}
\arguments{
\item{x}{:a vector}

\item{disc}{:discontinuity}

\item{disp}{:dispersion}
}
\description{
Compute the value of the unit normal loss integral, with discontinuity and dispersion
}
\examples{
 tr.unli(-3:10, 1, 3)
}
