% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xd.fred.tickers.R
\name{xd.fred.tickers}
\alias{xd.fred.tickers}
\title{Federal Reserve Bank of St. Louis Economic Data Tickers}
\usage{
xd.fred.tickers()
}
\description{
This function returns a data contains information of data name, type and tickers
}
\examples{
xd.fred.tickers()
}
